Build fix for Linux64 after r10529.
BUG=none
TEST=none
TBR=mnaganov

Review URL: https://chromiumcodereview.appspot.com/9225021

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10530 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/profile-generator.cc
@@ -2833,7 +2833,7 @@ class NativeGroupRetainedObjectInfo : public v8::RetainedObjectInfo {
 
  private:
   bool disposed_;
-  int hash_;
+  intptr_t hash_;
   const char* label_;
 };
